requirements.txt
requests>=2.25.1
paramiko>=2.7.2
mysql-connector-python>=8.0.23